# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "经济法知识图谱与上市公司治理合规推理系统 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/rules": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "List Rules",
        "operationId": "list_rules_api_v1_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleOut"
                  },
                  "type": "array",
                  "title": "Response List Rules Api V1 Rules Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/rules/{rule_id}": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Get Rule",
        "operationId": "get_rule_api_v1_rules__rule_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rules/{rule_id}/publish": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Publish Rule",
        "operationId": "publish_rule_api_v1_rules__rule_id__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleVersionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rulesets": {
      "get": {
        "tags": [
          "rulesets"
        ],
        "summary": "List Rulesets",
        "operationId": "list_rulesets_api_v1_rulesets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RuleSetStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleSetOut"
                  },
                  "title": "Response List Rulesets Api V1 Rulesets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rulesets"
        ],
        "summary": "Create Ruleset",
        "operationId": "create_ruleset_api_v1_rulesets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleSetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rulesets/{rule_set_id}": {
      "get": {
        "tags": [
          "rulesets"
        ],
        "summary": "Get Ruleset",
        "operationId": "get_ruleset_api_v1_rulesets__rule_set_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rulesets/{rule_set_id}/members": {
      "post": {
        "tags": [
          "rulesets"
        ],
        "summary": "Add Ruleset Member",
        "operationId": "add_ruleset_member_api_v1_rulesets__rule_set_id__members_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Set Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleSetMemberAdd"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rulesets/{rule_set_id}/publish": {
      "post": {
        "tags": [
          "rulesets"
        ],
        "summary": "Publish Ruleset",
        "operationId": "publish_ruleset_api_v1_rulesets__rule_set_id__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects": {
      "get": {
        "tags": [
          "subjects"
        ],
        "summary": "List Subjects",
        "operationId": "list_subjects_api_v1_subjects_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SubjectType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Type"
            }
          },
          {
            "name": "listed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Listed"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SubjectOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject_id}": {
      "get": {
        "tags": [
          "subjects"
        ],
        "summary": "Get Subject",
        "operationId": "get_subject_api_v1_subjects__subject_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subject Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subjects/{subject_id}/governance": {
      "get": {
        "tags": [
          "subjects"
        ],
        "summary": "Get Subject Governance",
        "operationId": "get_subject_governance_api_v1_subjects__subject_id__governance_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subject Id"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectGovernanceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrganizationGovernanceOut": {
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/OrganizationOut"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/RoleAssignmentOut"
            },
            "type": "array",
            "title": "Members"
          }
        },
        "type": "object",
        "required": [
          "organization",
          "members"
        ],
        "title": "OrganizationGovernanceOut"
      },
      "OrganizationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization_type": {
            "type": "string",
            "title": "Organization Type"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_type",
          "name"
        ],
        "title": "OrganizationOut"
      },
...
      "Page_SubjectOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SubjectOut"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "Page[SubjectOut]"
      },
...
      "RuleDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "latest_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuleVersionDetailOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "latest_version"
        ],
        "title": "RuleDetailOut"
      },
      "RuleOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "latest_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuleVersionSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "latest_version"
        ],
        "title": "RuleOut"
      },
      "RuleSetCreate": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "effective_from": {
            "type": "string",
            "format": "date",
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "effective_from"
        ],
        "title": "RuleSetCreate"
      },
      "RuleSetDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "version_no": {
            "type": "integer",
            "title": "Version No"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "effective_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/RuleVersionSummaryOut"
            },
            "type": "array",
            "title": "Members"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "version_no",
          "name",
          "status",
          "effective_from",
          "effective_to",
          "members"
        ],
        "title": "RuleSetDetailOut"
      },
      "RuleSetMemberAdd": {
        "properties": {
          "rule_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Rule Version Id"
          }
        },
        "type": "object",
        "required": [
          "rule_version_id"
        ],
        "title": "RuleSetMemberAdd"
      },
      "RuleSetOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "version_no": {
            "type": "integer",
            "title": "Version No"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "effective_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "version_no",
          "name",
          "status",
          "effective_from",
          "effective_to"
        ],
        "title": "RuleSetOut"
      },
      "RuleSetStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "PUBLISHED",
          "DEPRECATED"
        ],
        "title": "RuleSetStatus"
      },
...
      "RuleVersionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "version_no": {
            "type": "integer",
            "title": "Version No"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "version_no",
          "status"
        ],
        "title": "RuleVersionSummary"
      },
      "RuleVersionSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "rule_code": {
            "type": "string",
            "title": "Rule Code"
          },
          "version_no": {
            "type": "integer",
            "title": "Version No"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "rule_code",
          "version_no",
          "status"
        ],
        "title": "RuleVersionSummaryOut"
      },
      "SubjectGovernanceOut": {
        "properties": {
          "subject": {
            "$ref": "#/components/schemas/SubjectOut"
          },
          "at": {
            "type": "string",
            "format": "date",
            "title": "At"
          },
          "organizations": {
            "items": {
              "$ref": "#/components/schemas/OrganizationGovernanceOut"
            },
            "type": "array",
            "title": "Organizations"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "at",
          "organizations"
        ],
        "title": "SubjectGovernanceOut"
      },
      "SubjectOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "subject_type": {
            "type": "string",
            "title": "Subject Type"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "unified_credit_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unified Credit Code"
          },
          "listed": {
            "type": "boolean",
            "title": "Listed"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject_type",
          "name",
          "unified_credit_code",
          "listed",
          "exchange"
        ],
        "title": "SubjectOut"
      },
      "SubjectType": {
        "type": "string",
        "enum": [
          "COMPANY",
          "LISTED_COMPANY",
          "PERSON",
          "REGULATOR"
        ],
        "title": "SubjectType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}